+++ CHANGELOG.md
@@ -1,9 +1,12 @@
 # Changelog
 
+## 0.1.2 (2015-04-01)
+Features:
+ + Adds fabutils.context.CommandMessage class to squash command output with fancy messages
+
 ## 0.1.1 (2015-03-06)
 Features:
-    * Adds validation againts json schema
+ + Adds validation against json schema
 
 ## 0.1.0 (2015-03-04)
-
 Initial Release